_option.rswitch removed,Bus_3ph P,Q results fixed #96

+++ pandapower/auxiliary.py
@@ -757,7 +757,7 @@ def SVabc_from_SV012(S012, V012, n_res=None, idx=None):
         n_res = S012.shape[1]
     if idx is None:
         idx = np.ones(n_res, dtype="bool")
-    I012 = np.array(np.zeros((3, n_res)))
+    I012 = np.array(np.zeros((3, n_res)),dtype = np.complex128)
     I012[:, idx] = I_from_SV_elementwise(S012[:, idx], V012[:, idx])
     Vabc = sequence_to_phase(V012[:, idx])
     Iabc = sequence_to_phase(I012[:, idx])
